+++ src/spartan/polys/eq.rs
